Created j/hardware/easyKeys
A tool to modify the output of a MIDI keyboard so that the white keys can
be mapped to any Major or Minor key using the black keys as the controls.
The F# black key can be used to output diatonic triad chords instead of
just single notes.

Created j/convert/numbers2symbols

A list of symbols to the right inlet can be called by an index at the left
input. This takes an optional argument of a canvas name to send the symbol
as the canvas's label for GUI purposes.